split up control from matrix file and moved led.c helper functions into the LEDCloudMatrix namespace
